Move docx conversion fns to Util file
