+++ src/resolvers/Package.js
@@ -1,11 +1,8 @@
 import { ApolloError, UserInputError } from 'apollo-server-express';
-import axios from 'axios';
-import mammoth from 'mammoth';
-import toArray from 'stream-to-array';
-import xss from 'xss';
 // import pubsub from '../services/pubsub';
+import { convertDocxContent } from './Util';
 import {
-  deleteAllS3Assets, deleteS3Asset, getAssetPath, getSignedUrlPromiseGet
+  deleteAllS3Assets, deleteS3Asset, getAssetPath
 } from '../services/aws/s3';
 // import transformPackage from '../services/es/package/transform';
 // import { publishCreate, publishUpdate, publishDelete } from '../services/rabbitmq/package';
@@ -13,108 +10,6 @@ import { hasValidValue } from '../lib/projectParser';
 import { PACKAGE_FULL } from '../fragments/package';
 
 const PUBLISHER_BUCKET = process.env.AWS_S3_AUTHORING_BUCKET;
-
-/**
- * Strips tags and multiple spaces from a string of html
- * @param {string} string
- */
-const htmlToText = ( string = '' ) => (
-  string
-    .replace( /<[\s\S]*?>/g, ' ' )
-    .replace( /\s{2,}/g, ' ' )
-    .trim()
-);
-
-/**
- * Creates a Buffer from a stream part
- * @param {*} part
- */
-const handleStreamPart = part => (
-  Buffer.isBuffer( part )
-    ? part
-    : Buffer.from( part )
-);
-
-/**
- * Concatenates an array of Buffers
- * @param {array} streamParts
- */
-const streamToBuffer = streamParts => {
-  const buffers = streamParts.map( handleStreamPart );
-  return Buffer.concat( buffers );
-};
-
-/**
- * Gets a remote docx as a Buffer
- * @param {string} url
- */
-const getDocxBuffer = url => (
-  axios.get( url, { responseType: 'stream' } )
-    .then( async res => {
-      if ( res.status === 200 ) {
-        return toArray( res.data )
-          .then( streamToBuffer )
-          .then( buf => buf ) // buf, i.e., buffer
-          .catch( err => console.log( err ) );
-      }
-      return [];
-    } )
-    .then( buf => buf )
-    .catch( err => console.log( err ) )
-);
-
-/**
- * Builds document.content w/ converted docx content
- * @param {object} document
- * @param {object} input
- * @see https://github.com/mwilliamson/mammoth.js
- */
-const setDocumentContent = async ( document, input ) => (
-  mammoth.convertToHtml( input )
-    .then( result => {
-      if ( result.value ) {
-        document.content = {
-          create: {
-            rawText: htmlToText( result.value ),
-            html: xss( result.value ),
-            // omitting md to simply & since it's not used by client
-            markdown: ''
-          }
-        };
-      }
-
-      if ( result.messages && result.messages.length ) {
-        result.messages.forEach( msg => console.log( msg ) );
-      }
-    } )
-    .catch( err => console.log( err ) )
-    .done()
-);
-
-/**
- * Converts docx
- * @param {string} id Package id
- * @param {object} data
- * @param {object} ctx Prisma context
- */
-const convertDocxContent = async ( id, data, ctx ) => {
-  if ( !data.documents || !data.documents.create || !data.documents.create[0] ) {
-    return;
-  }
-
-  const document = data.documents.create[0];
-  const signed = await getSignedUrlPromiseGet( { key: document.url } );
-  const buffer = await getDocxBuffer( signed.url );
-
-  await setDocumentContent( document, { buffer } )
-    .then( () => (
-      // call updatePackage to update data w/ document content
-      ctx.prisma.updatePackage( {
-        data: {}, // {} to avoid creating a document w/ no content
-        where: { id }
-      } )
-    ) );
-};
 
 export default {
   Subscription: {},
